Refactor indexing modules per source type
