+++ pycontextify/index.py
@@ -5,14 +5,24 @@ embedding providers, vector storage, content loading, text chunking,
 and lightweight knowledge graph capabilities.
 """
 
+from .index_codebase import CodeLoader, CodebaseIndexer
+from .index_document import DocumentIndexer, DocumentLoader, PDFLoader
+from .index_webpage import WebpageIndexer, WebpageLoader
 from .indexer_manager import IndexManager
 from .config import Config
 from .storage_metadata import ChunkMetadata, MetadataStore, SourceType
 
 __all__ = [
+    "CodeLoader",
+    "CodebaseIndexer",
     "IndexManager",
     "Config",
     "ChunkMetadata",
     "SourceType",
     "MetadataStore",
+    "DocumentIndexer",
+    "DocumentLoader",
+    "PDFLoader",
+    "WebpageIndexer",
+    "WebpageLoader",
 ]
