+++ tests/integration/test_loaders.py
@@ -9,7 +9,8 @@ pytest.importorskip("crawl4ai")
 
 from crawl4ai.models import CrawlResult, MarkdownGenerationResult
 
-from pycontextify.indexer_loaders import CodeLoader, WebpageLoader
+from pycontextify.index_codebase import CodeLoader
+from pycontextify.index_webpage import WebpageLoader
 
 
 def _mk_crawl_result(url: str, text: str) -> CrawlResult:
